Re-organize classes to folders
