+++ app/src/main/java/com/udacity/popularmovies/ui/MovieTrailerAdapter.java
@@ -1,6 +1,8 @@
-package com.udacity.popularmovies;
+package com.udacity.popularmovies.ui;
 
 import android.content.Context;
+import android.content.Intent;
+import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,62 +12,62 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.udacity.popularmovies.model.ReviewDetails;
+import com.udacity.popularmovies.utilies.App;
+import com.udacity.popularmovies.R;
+import com.udacity.popularmovies.model.TrailerDetails;
 import com.udacity.popularmovies.network.AssertConnectivity;
+
 
 import java.util.ArrayList;
 
-public class MovieReviewAdapter extends RecyclerView.Adapter<MovieReviewAdapter.ViewHolder>{
+// TODO: SEARCH IF MULTIPLE RECYCLERVIEW IS NORMAL IMPLEMENTATION
+public class MovieTrailerAdapter extends RecyclerView.Adapter<MovieTrailerAdapter.ViewHolder> {
 
     private Context mContext;
-    private ArrayList<ReviewDetails> review;
+    private ArrayList<TrailerDetails> trailer;
 
-    public MovieReviewAdapter(Context mContext, ArrayList<ReviewDetails> review) {
+    public MovieTrailerAdapter(Context mContext, ArrayList<TrailerDetails> trailer){
         this.mContext = mContext;
-        this.review = review;
+        this.trailer = trailer;
     }
 
+
+    // Required for RecyclerView. Responsible for inflating the view / recycling the viewholder
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.layout_movie_review, parent, false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.layout_movie_trailer, parent, false);
         ViewHolder holder = new ViewHolder(view);
         return holder;
     }
 
+    // Required for RecyclerView. Changes depends on what layouts are
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        holder.reviewAuthor.setText(review.get(position).getAuthor());
-        holder.reviewContent.setText(review.get(position).getContent());
-
+        holder.trailerTitle.setText(trailer.get(position).getName());
 
     }
 
     @Override
-    public int getItemCount() { return review != null? review.size() : 0; }
-
+    public int getItemCount() { return trailer != null? trailer.size() : 0; }
 
 
     // Holds widget in memory for each individual entry
     public class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
 
-        TextView reviewAuthor;
-        TextView reviewContent;
-
+        TextView trailerTitle;
         RelativeLayout parentLayout;
 
         //Constructor required for Viewholder
         public ViewHolder( View itemView) {
             super(itemView);
-            reviewAuthor = itemView.findViewById(R.id.text_reviewAuthor);
-            reviewContent = itemView.findViewById(R.id.text_reviewContent);
-
-            parentLayout = itemView.findViewById(R.id.reviewParent_layout);
+            trailerTitle = itemView.findViewById(R.id.text_trailerTitle);
+            parentLayout = itemView.findViewById(R.id.trailerParent_layout);
             itemView.setOnClickListener(this);
         }
 
 
-        // TODO: CHECK IF SHOULD BE ADDED
+
         @Override
         public void onClick(View v) {
 
@@ -74,10 +76,10 @@ public class MovieReviewAdapter extends RecyclerView.Adapter<MovieReviewAdapter.
 
             if(AssertConnectivity.isOnline()){
                 int position = getAdapterPosition();
-/*                // Get URL and open the trailer link
+                // Get URL and open the trailer link
                 Uri uri = Uri.parse(trailer.get(position).getKey());
                 Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-                mContext.startActivity(intent);*/
+                mContext.startActivity(intent);
             }else
                 AssertConnectivity.errorConnectMessage(App.getAppResources().getString(R.string.error_connection_general));
 
@@ -87,6 +89,7 @@ public class MovieReviewAdapter extends RecyclerView.Adapter<MovieReviewAdapter.
 
         }
     }
+
 
 
 }
